requirements.txt
streamlit
langchain
langgraph
chromadb
openai
tiktoken
plotly
scikit-learn
python-dotenv
nltk